+++ internal/account/model/borrow.go
@@ -11,8 +11,8 @@ type BorrowLendFlow struct {
 	Type       int     `json:"type"`
 	RecordTime int64   `json:"record_time"`
 	// 账户id
-	AccountId int `json:"account_id"`
+	AccountId int64 `json:"account_id"`
 	// 关联的借入借出id
-	BorrowId int    `json:"borrow_id"`
+	BorrowId int64  `json:"borrow_id"`
 	Remark   string `json:"remark"`
 }
